refactor(All): 패키지 구조 변경
- domain 패키지 안에 model 패키지를 둘 필요가 없다. model 패키지를 삭제한다.
