+++ src/main/java/attendance/controller/AttendanceController.java
@@ -5,8 +5,8 @@ import attendance.controller.command.Command;
 import attendance.controller.command.FindByCrewCommand;
 import attendance.controller.command.FindDismissalCommand;
 import attendance.controller.command.ModifyCommand;
-import attendance.domain.model.Campus;
-import attendance.domain.model.CrewHistories;
+import attendance.domain.Campus;
+import attendance.domain.CrewHistories;
 import attendance.view.InputView;
 import attendance.view.MenuOption;
 import attendance.view.ResultView;
