+++ src/main/java/attendance/controller/command/Command.java
@@ -1,6 +1,6 @@
 package attendance.controller.command;
 
-import attendance.domain.model.CrewHistories;
+import attendance.domain.CrewHistories;
 
 public interface Command {
