+++ src/main/java/attendance/controller/command/FindDismissalCommand.java
@@ -1,7 +1,7 @@
 package attendance.controller.command;
 
-import attendance.domain.model.AttendanceCounter;
-import attendance.domain.model.CrewHistories;
+import attendance.domain.AttendanceCounter;
+import attendance.domain.CrewHistories;
 import attendance.dto.AttendanceWarning;
 import attendance.view.ResultView;
 import java.time.Clock;
