+++ src/main/java/attendance/domain/AttendanceStatus.java
@@ -1,4 +1,4 @@
-package attendance.domain.model;
+package attendance.domain;
 
 import java.time.DayOfWeek;
 import java.time.Duration;
