+++ src/main/java/attendance/domain/WarningLevel.java
@@ -1,4 +1,4 @@
-package attendance.domain.model;
+package attendance.domain;
 
 import java.util.Arrays;
 import java.util.Comparator;
