+++ src/main/java/attendance/dto/AttendanceWarning.java
@@ -3,8 +3,8 @@ package attendance.dto;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import attendance.domain.model.AttendanceCounter;
-import attendance.domain.model.WarningLevel;
+import attendance.domain.AttendanceCounter;
+import attendance.domain.WarningLevel;
 
 public record AttendanceWarning(String nickname, int absentCount, int lateCount, WarningLevel warningLevel) {
