+++ src/main/java/attendance/util/TimeFormatter.java
@@ -1,6 +1,6 @@
 package attendance.util;
 
-import static attendance.domain.model.AttendanceStatus.DEFAULT_TIME;
+import static attendance.domain.AttendanceStatus.DEFAULT_TIME;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
