+++ src/test/java/attendance/fixture/TestFixture.java
@@ -1,6 +1,6 @@
 package attendance.fixture;
 
-import static attendance.domain.model.AttendanceStatus.DEFAULT_TIME;
+import static attendance.domain.AttendanceStatus.DEFAULT_TIME;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
